bugfix: change short by int

+++ streaming/src/main/scala/com/mitosis/beans/FlightInfoBean.scala
@@ -20,7 +20,7 @@ class FlightInfoBean {
   var arrivingDate: Long = _
   
   @BeanProperty
-  var passengerNumber: Short = _
+  var passengerNumber: Int = _
   
   @BeanProperty
   var cabinClass: String = _
